+++ sdk/darfPayment/log/index.js
@@ -1,0 +1,6 @@
+const log = require('./log.js');
+
+exports.get = log.get;
+exports.query = log.query;
+exports.page = log.page;
+exports.Log = log.Log;