+++ tests/testDarfPayment.js
@@ -1,0 +1,74 @@
+const assert = require('assert');
+const starkbank = require('../index.js');
+const generateExampleDarfPaymentsJson = require('./utils/taxPayment').generateExampleDarfPaymentsJson;
+
+starkbank.user = require('./utils/user').exampleProject;
+
+
+describe('TestDarfPaymentPost', function(){
+    this.timeout(10000);
+    it('test_success', async () => {
+        let payments = await generateExampleDarfPaymentsJson(5);
+        payments = await starkbank.darfPayment.create(payments);
+        for (let payment of payments) {
+            assert(typeof payment.id == 'string');
+        }
+    });
+});
+
+describe('TestDarfPaymentGetPage', function () {
+    this.timeout(10000);
+    it('test_success', async () => {
+        let ids = [];
+        let cursor = null;
+        let page = null;
+        for (let i = 0; i < 2; i++) {
+            [page, cursor] = await starkbank.darfPayment.page({ limit: 5, cursor: cursor });
+            for (let entity of page) {
+                assert(!ids.includes(entity.id));
+                ids.push(entity.id);
+            }
+            if (cursor == null) {
+                break;
+            }
+        }
+        assert(ids.length == 10);
+    });
+});
+
+describe('TestDarfPaymentGet', function(){
+    this.timeout(10000);
+    it('test_success', async () => {
+        let i = 0;
+        const payments = await starkbank.darfPayment.query({limit: 5});
+        for await (let payment of payments) {
+            assert(typeof payment.id == 'string');
+            i += 1;
+        }
+        assert(i === 5);
+    });
+});
+
+describe('TestDarfPaymentPdfGet', function(){
+    this.timeout(10000);
+    it('test_success', async () => {
+        let payments = await starkbank.darfPayment.query({limit: 1, status: 'processing'});
+        for await (let payment of payments) {
+            assert(typeof payment.id == 'string');
+            let pdf = await starkbank.darfPayment.pdf(payment.id);
+            assert(Buffer.isBuffer(pdf));
+        }
+    });
+});
+
+describe('TestDarfPaymentPostAndDelete', function(){
+    this.timeout(10000);
+    it('test_success', async () => {
+        let payments = await generateExampleDarfPaymentsJson(1, true);
+        payments = await starkbank.darfPayment.create(payments);
+        let payment = payments[0];
+        assert(typeof payment.id == 'string');
+        payment = await starkbank.darfPayment.delete(payment.id);
+        assert(typeof payment.id == 'string');
+    });
+});